+++ packages/core/src/RenderingEngine/VideoViewport.ts
@@ -1,4 +1,9 @@
-import { Events as EVENTS, VideoViewport as VideoViewportEnum } from '../enums';
+import { vec3 } from 'gl-matrix';
+import {
+  Events as EVENTS,
+  VideoViewport as VideoViewportEnum,
+  MetadataModules,
+} from '../enums';
 import {
   IVideoViewport,
   VideoViewportProperties,
@@ -7,7 +12,9 @@ import {
   ICamera,
   InternalVideoCamera,
   VideoViewportInput,
+  VOIRange,
 } from '../types';
+import * as metaData from '../metaData';
 import { Transform } from './helpers/cpuFallback/rendering/transform';
 import { triggerEvent } from '../utilities';
 import Viewport from './Viewport';
@@ -18,7 +25,9 @@ import { getOrCreateCanvas } from './helpers';
  * looking into an internal scene, and an associated target output `canvas`.
  */
 class VideoViewport extends Viewport implements IVideoViewport {
+  public modality;
   // Viewport Data
+  protected imageId: string;
   readonly uid;
   readonly renderingEngineId: string;
   readonly canvasContext: CanvasRenderingContext2D;
@@ -30,10 +39,41 @@ class VideoViewport extends Viewport implements IVideoViewport {
   private mute = true;
   private isPlaying = false;
   private scrollSpeed = 1;
-  private fps = 30; // TODO We need to find a good solution for this.
+  private playbackRate = 1;
+
+  protected metadata;
+
+  /**
+   * The fps, frames per second is used to calculate time/frame mapping values.
+   * It is provided by the CINE Module in the metadata, defaulting to 30 if not
+   * provided.
+   */
+  private fps = 30;
+
   private videoCamera: InternalVideoCamera = {
     panWorld: [0, 0],
     parallelScale: 1,
+  };
+
+  /**
+   * feFilter is an inline string value for the CSS filter on the video
+   * CSS filters can reference SVG filters, so for the typical use case here
+   * the CSS filter is actually an link link to a SVG filter.
+   */
+  private feFilter: string;
+
+  /**
+   * An average white point value, used to color balance the image so that
+   * the given white is mapped to [255,255,255] via multiplication per channel.
+   */
+  private averageWhite: [number, number, number];
+
+  /**
+   * The VOI Range is used to apply contrast/brightness adjustments to the image.
+   */
+  private voiRange: VOIRange = {
+    lower: 0,
+    upper: 255,
   };
 
   constructor(props: VideoViewportInput) {
@@ -82,6 +122,86 @@ class VideoViewport extends Viewport implements IVideoViewport {
     this.videoElement.remove();
   }
 
+  private _getImageDataMetadata() {
+    const imagePlaneModule = metaData.get(
+      MetadataModules.IMAGE_PLANE,
+      this.imageId
+    );
+
+    let rowCosines = <Point3>imagePlaneModule.rowCosines;
+    let columnCosines = <Point3>imagePlaneModule.columnCosines;
+
+    // if null or undefined
+    if (rowCosines == null || columnCosines == null) {
+      rowCosines = <Point3>[1, 0, 0];
+      columnCosines = <Point3>[0, 1, 0];
+    }
+
+    const rowCosineVec = vec3.fromValues(
+      rowCosines[0],
+      rowCosines[1],
+      rowCosines[2]
+    );
+    const colCosineVec = vec3.fromValues(
+      columnCosines[0],
+      columnCosines[1],
+      columnCosines[2]
+    );
+    const scanAxisNormal = vec3.create();
+    vec3.cross(scanAxisNormal, rowCosineVec, colCosineVec);
+
+    let origin = imagePlaneModule.imagePositionPatient;
+    // if null or undefined
+    if (origin == null) {
+      origin = [0, 0, 0];
+    }
+
+    const xSpacing = imagePlaneModule.columnPixelSpacing;
+    const ySpacing = imagePlaneModule.rowPixelSpacing;
+    const xVoxels = imagePlaneModule.columns;
+    const yVoxels = imagePlaneModule.rows;
+
+    const zSpacing = 1;
+    const zVoxels = 1;
+
+    return {
+      bitsAllocated: 8,
+      numComps: 3,
+      origin,
+      direction: [...rowCosineVec, ...colCosineVec, ...scanAxisNormal],
+      dimensions: [xVoxels, yVoxels, zVoxels],
+      spacing: [xSpacing, ySpacing, zSpacing],
+      numVoxels: xVoxels * yVoxels * zVoxels,
+      imagePlaneModule,
+    };
+  }
+
+  /**
+   * Sets the video image id to show and hte frame number.
+   * Requirements are to have the imageUrlModule in the metadata
+   * with the rendered endpoint being the raw video in video/mp4 format.
+   */
+  public setVideo(
+    imageIds: string | string[],
+    frameNumber?: number
+  ): Promise<unknown> {
+    this.imageId = Array.isArray(imageIds) ? imageIds[0] : imageIds;
+    const { imageId } = this;
+    const { rendered } = metaData.get(MetadataModules.IMAGE_URL, imageId);
+    const generalSeries = metaData.get(MetadataModules.GENERAL_SERIES, imageId);
+    this.modality = generalSeries?.Modality;
+    this.metadata = this._getImageDataMetadata();
+
+    return this.setVideoURL(rendered).then(() => {
+      const { cineRate = 30 } = metaData.get(MetadataModules.CINE, imageId);
+      this.fps = cineRate;
+      if (frameNumber !== undefined) {
+        this.pause();
+        this.setFrame(frameNumber);
+      }
+    });
+  }
+
   public async setVideoURL(videoURL: string) {
     return new Promise((resolve) => {
       this.videoElement.src = videoURL;
@@ -126,10 +246,8 @@ class VideoViewport extends Viewport implements IVideoViewport {
   }
 
   public async pause() {
-    if (this.isPlaying) {
-      await this.videoElement.pause();
-      this.isPlaying = false;
-    }
+    await this.videoElement.pause();
+    this.isPlaying = false;
   }
 
   public async scroll(delta = 1) {
@@ -162,8 +280,6 @@ class VideoViewport extends Viewport implements IVideoViewport {
     if (videoElement.paused) {
       // Need to wait for seek update
       const seekEventListener = (evt) => {
-        console.log('seeked');
-
         renderFrame();
 
         videoElement.removeEventListener('seeked', seekEventListener);
@@ -214,21 +330,30 @@ class VideoViewport extends Viewport implements IVideoViewport {
     this.setTime((frame - 1) / this.fps);
   }
 
-  public setProperties(videoInterface: VideoViewportProperties) {
-    if (videoInterface.loop !== undefined) {
-      this.videoElement.loop = videoInterface.loop;
+  public setProperties(props: VideoViewportProperties) {
+    if (props.loop !== undefined) {
+      this.videoElement.loop = props.loop;
     }
 
-    if (videoInterface.muted !== undefined) {
-      this.videoElement.muted = videoInterface.muted;
+    if (props.muted !== undefined) {
+      this.videoElement.muted = props.muted;
     }
 
-    if (videoInterface.playbackRate !== undefined) {
-      this.setPlaybackRate(videoInterface.playbackRate);
+    if (props.playbackRate !== undefined) {
+      this.setPlaybackRate(props.playbackRate);
+    }
+
+    if (props.scrollSpeed !== undefined) {
+      this.setScrollSpeed(props.scrollSpeed);
+    }
+
+    if (props.voiRange) {
+      this.setVOI(props.voiRange);
     }
   }
 
   public setPlaybackRate(rate = 1) {
+    this.playbackRate = rate;
     // Minimum playback speed in chrome is 0.0625 compared to normal
     if (rate < 0.0625) {
       this.pause();
@@ -255,6 +380,9 @@ class VideoViewport extends Viewport implements IVideoViewport {
     return {
       loop: this.videoElement.loop,
       muted: this.videoElement.muted,
+      playbackRate: this.playbackRate,
+      scrollSpeed: this.scrollSpeed,
+      voiRange: { ...this.voiRange },
     };
   };
 
@@ -265,8 +393,101 @@ class VideoViewport extends Viewport implements IVideoViewport {
     });
   }
 
+  protected getScalarData() {
+    const canvas = document.createElement('canvas');
+    canvas.width = this.videoWidth;
+    canvas.height = this.videoHeight;
+    const context = canvas.getContext('2d');
+    context.drawImage(this.videoElement, 0, 0);
+    const canvasData = context.getImageData(
+      0,
+      0,
+      this.videoWidth,
+      this.videoHeight
+    );
+    const { data: scalarData } = canvasData;
+    (scalarData as any).getRange = () => [0, 255];
+    return scalarData;
+  }
+
   public getImageData() {
-    return null;
+    const { metadata } = this;
+
+    const spacing = metadata.spacing;
+
+    return {
+      dimensions: metadata.dimensions,
+      spacing,
+      origin: metadata.origin,
+      direction: metadata.direction,
+      metadata: { Modality: this.modality },
+      imageData: {
+        getDirection: () => metadata.direction,
+        getDimensions: () => metadata.dimensions,
+        getRange: () => [0, 255],
+        getScalarData: () => this.getScalarData(),
+        getSpacing: () => metadata.spacing,
+        worldToIndex: (point: Point3) => {
+          const canvasPoint = this.worldToCanvas(point);
+          const pixelCoord = this.canvasToIndex(canvasPoint);
+          return [pixelCoord[0], pixelCoord[1], 0];
+        },
+        indexToWorld: (point: Point3) => {
+          const canvasPoint = this.indexToCanvas([point[0], point[1]]);
+          return this.canvasToWorld(canvasPoint);
+        },
+      },
+      hasPixelSpacing: this.hasPixelSpacing,
+      calibration: this.calibration,
+      preScale: {
+        scaled: false,
+      },
+    };
+  }
+
+  public setVOI(voiRange: VOIRange): void {
+    this.voiRange = voiRange;
+    this.setColorTransform();
+  }
+
+  public setWindowLevel(windowWidth = 256, windowCenter = 128) {
+    const lower = windowCenter - windowWidth / 2;
+    const upper = windowCenter + windowWidth / 2 - 1;
+    this.setVOI({ lower, upper });
+    this.setColorTransform();
+  }
+
+  public setAverageWhite(averageWhite: [number, number, number]) {
+    this.averageWhite = averageWhite;
+    this.setColorTransform();
+  }
+
+  protected setColorTransform() {
+    if (!this.voiRange && !this.averageWhite) {
+      this.feFilter = null;
+      return;
+    }
+    const white = this.averageWhite || [255, 255, 255];
+    const maxWhite = Math.max(...white);
+    const scaleWhite = white.map((c) => maxWhite / c);
+    // From the DICOM standard: ((x - (c - 0.5)) / (w-1) + 0.5) * (ymax- ymin) + ymin
+    // which is x/(w-1) - (c - 0.5) / (w-1) + 0.5  for this case
+    const { lower = 0, upper = 255 } = this.voiRange || {};
+    const wlScale = (upper - lower + 1) / 255;
+    const wlDelta = lower / 255;
+    this.feFilter = `url('data:image/svg+xml,\
+      <svg xmlns="http://www.w3.org/2000/svg">\
+        <filter id="colour" color-interpolation-filters="linearRGB">\
+        <feColorMatrix type="matrix" \
+        values="\
+          ${scaleWhite[0] * wlScale} 0 0 0 ${wlDelta} \
+          0 ${scaleWhite[1] * wlScale} 0 0 ${wlDelta} \
+          0 0 ${scaleWhite[2] * wlScale} 0 ${wlDelta} \
+          0 0 0 1 0" />\
+        </filter>\
+      </svg>#colour')`;
+
+    this.canvas.style.filter = this.feFilter;
   }
 
   public setCamera(camera: ICamera): void {
@@ -339,6 +560,10 @@ class VideoViewport extends Viewport implements IVideoViewport {
     return true;
   };
 
+  public getNumberOfSlices = (): number => {
+    return (this.videoElement.duration * this.fps) / this.scrollSpeed;
+  };
+
   public getFrameOfReferenceUID = (): string => {
     // The video itself is the frame of reference.
     return this.videoElement.src;
@@ -409,6 +634,20 @@ class VideoViewport extends Viewport implements IVideoViewport {
     const canvasPos: Point2 = [subCanvasPos[0], subCanvasPos[1]];
 
     return canvasPos;
+  };
+
+  protected canvasToIndex = (canvasPos: Point2): Point2 => {
+    const [x, y] = canvasPos;
+    const ratio = this.videoWidth / this.canvas.width;
+    const pan = this.getPan();
+    return [(x + pan[0]) * ratio, (y + pan[1]) * ratio];
+  };
+
+  protected indexToCanvas = (indexPos: Point2): Point2 => {
+    const [x, y] = indexPos;
+    const ratio = this.canvas.width / this.videoWidth;
+    const pan = this.getPan();
+    return [x * ratio - pan[0], y * ratio - pan[1]];
   };
 
   private refreshRenderValues() {