+++ packages/core/src/types/VideoViewportProperties.ts
@@ -9,9 +9,7 @@ type VideoViewportProperties = ViewportProperties & {
   muted?: boolean;
   pan?: Point2;
   playbackRate?: number;
-  // The zoom factor, naming consistent with vtk cameras for now,
-  // but this isn't necessarily necessary.
-  parallelScale?: number;
+  scrollSpeed?: number;
 };
 
 export default VideoViewportProperties;