+++ packages/streaming-image-volume-loader/src/StreamingImageVolume.ts
@@ -36,8 +36,13 @@ export default class StreamingImageVolume extends BaseStreamingImageVolume {
    */
   public getImageLoadRequests(priority: number): ImageLoadRequests[] {
     const { imageIds } = this;
-    const scalarData = <Types.VolumeScalarData>this.scalarData;
 
-    return this.getImageIdsRequests(imageIds, scalarData, priority);
+    return this.getImageIdsRequests(imageIds, priority);
   }
+
+  public getImageIdsToLoad = () => {
+    const { imageIds } = this;
+    this.numFrames = imageIds.length;
+    return imageIds;
+  };
 }