+++ packages/streaming-image-volume-loader/src/cornerstoneStreamingImageVolumeLoader.ts
@@ -41,6 +41,7 @@ function cornerstoneStreamingImageVolumeLoader(
   volumeId: string,
   options: {
     imageIds: string[];
+    progressiveRendering?: boolean | Types.IRetrieveConfiguration;
   }
 ): IVolumeLoader {
   if (!options || !options.imageIds || !options.imageIds.length) {
@@ -92,7 +93,7 @@ function cornerstoneStreamingImageVolumeLoader(
       ).catch(console.error);
     }
 
-    const { imageIds } = options;
+    const { imageIds, progressiveRendering } = options;
 
     const volumeMetadata = csUtils.makeVolumeMetadata(imageIds);